Accounting for empty channels
Some files may have waveforms dropped from one or two channels. If this happens in the first channel in the tetrode the variables: curValley, curValleyINds, curMinValley and curMinValleyNdx were not being made with the correct index size meaning with each iteration those variables were being expanded. Adding in this section makes it so they are now appropriately made if one of the channels turns out to have waveforms but the others don't.